feat(tooling): move webpack to its own folder
