+++ .webpack/webpack.common.js
@@ -20,14 +20,37 @@
  * at runtime from the About dialog for additional information.
  *****************************************************************************/
 
-const { merge } = require('webpack-merge');
-const common = require('./webpack.common');
+const path = require('path');
+const projectRootDir = path.resolve(__dirname, '..');
+
+// eslint-disable no-undef
+const WEBPACK_COMMON_CONFIG = {
+    performance: {
+        hints: false
+    },
+    resolve: {
+        alias: {
+            saveAs: "file-saver/src/FileSaver.js",
+        }
+    },
+    module: {
+        rules: [
+            {
+                test: /\.js$/,
+                enforce: "pre",
+                use: ["source-map-loader"]
+            }
+        ]
+    },
+    output: {
+        globalObject: "this",
+        filename: '[name].js',
+        // eslint-disable-next-line no-undef
+        path: path.resolve(projectRootDir, 'dist'),
+        libraryTarget: 'umd',
+        library: 'openmctYamcs'
+    }
+};
 
 // eslint-disable-next-line no-undef
-module.exports = merge(common, {
-    mode: 'production',
-    entry: {
-        'openmct-yamcs': './src/plugin.js'
-    },
-    devtool: 'source-map'
-});
+module.exports = WEBPACK_COMMON_CONFIG;
