+++ .webpack/webpack.prod.js
@@ -21,35 +21,16 @@
  *****************************************************************************/
 
 const path = require('path');
+const { merge } = require('webpack-merge');
+const common = require('./webpack.common');
 
-// eslint-disable no-undef
-const WEBPACK_COMMON_CONFIG = {
-    performance: {
-        hints: false
-    },
-    resolve: {
-        alias: {
-            saveAs: "file-saver/src/FileSaver.js",
-        }
-    },
-    module: {
-        rules: [
-            {
-                test: /\.js$/,
-                enforce: "pre",
-                use: ["source-map-loader"]
-            }
-        ]
-    },
-    output: {
-        globalObject: "this",
-        filename: '[name].js',
-        // eslint-disable-next-line no-undef
-        path: path.resolve(__dirname, 'dist'),
-        libraryTarget: 'umd',
-        library: 'openmctYamcs'
-    }
-};
-
+const projectRootDir = path.resolve(__dirname, '..');
 // eslint-disable-next-line no-undef
-module.exports = WEBPACK_COMMON_CONFIG;
+module.exports = merge(common, {
+    context: projectRootDir,
+    mode: 'production',
+    entry: {
+        'openmct-yamcs': path.resolve(projectRootDir, 'src/plugin.js')
+    },
+    devtool: 'source-map'
+});
