+++ package.json
@@ -8,12 +8,12 @@
     "test": "npm run test:getopensource && npm run test:e2e:smoke",
     "lint": "eslint src example",
     "lint:fix": "eslint src example --fix",
-    "build:dist": "webpack --config webpack.prod.js",
+    "build:dist": "webpack --config ./.webpack/webpack.prod.js",
     "build:example": "npm install openmct@unstable --no-save",
     "build:example:master": "npm install nasa/openmct#master --no-save",
     "postbuild:example": "node check-optional-dependencies.js",
-    "start": "npx webpack serve --config ./webpack.dev.js",
-    "start:coverage": "npx webpack serve --config ./webpack.coverage.js",
+    "start": "npx webpack serve --config ./.webpack/webpack.dev.js",
+    "start:coverage": "npx webpack serve --config ./.webpack/webpack.coverage.js",
     "prepare": "npm run build:dist",
     "test:getopensource": "sh ./tests/git-opensource-tests.sh",
     "test:e2e:smoke": "npx playwright test --config=./tests/e2e/playwright-quickstart.config.js --project=chromium quickstartSmoke",
